requirements.txt
Django==2.2.19
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
requests==2.26.0
pytest-pythonpath==0.7.3
django-filter==21.1
